seperate app_config and soundboards in two modules
