+++ Cargo.toml
@@ -40,6 +40,7 @@ crossbeam-channel = "0.4"
 dirs = "3"
 env_logger = "0.7"
 hotkey-soundboard = {path = "extern/hotkey-rs", version = "0.0.2"}
+indexmap = "1.5"
 log = "0.4"
 mime_guess = "2.0.3"
 miniaudio = "0.10.0"
