+++ src/soundboards.rs
@@ -22,88 +22,16 @@ mod helpers;
 
 use helpers::*;
 
-type GlobalAppConfig = Lazy<parking_lot::RwLock<std::sync::Arc<AppConfig>>>;
-
 pub type SoundboardId = Ulid;
-type SoundboardMap = HashMap<SoundboardId, Soundboard>;
+type SoundboardMap = indexmap::IndexMap<SoundboardId, Soundboard>;
 pub type SoundId = Ulid;
 type SoundMap = HashMap<SoundId, Sound>;
 type GlobalSoundboardMap = Lazy<parking_lot::RwLock<std::sync::Arc<SoundboardMap>>>;
-
-static GLOBAL_APP_CONFIG: GlobalAppConfig = Lazy::new(|| {
-    let app_config = load_and_merge_app_config().expect("failed to load and merge app config");
-    parking_lot::RwLock::new(std::sync::Arc::new(app_config))
-});
 
 static GLOBAL_SOUNDBOARD_MAP: GlobalSoundboardMap = Lazy::new(|| {
     let soundboards = load_and_parse_soundboards().expect("failed to load soundboards");
     parking_lot::RwLock::new(std::sync::Arc::new(soundboards))
 });
-
-#[derive(Debug, Deserialize, Default, Clone, Serialize)]
-pub struct AppConfig {
-    pub input_device: Option<String>,
-    pub output_device: Option<String>,
-    pub loopback_device: Option<String>,
-    pub stop_hotkey: Option<String>,
-    pub http_socket_addr: Option<String>,
-
-    pub no_http_server: Option<bool>,
-    pub telegram: Option<bool>,
-    pub terminal_ui: Option<bool>,
-    pub no_native_gui: Option<bool>,
-    pub auto_loop_device: Option<bool>,
-    pub no_duplex_device: Option<bool>,
-    pub print_possible_devices: Option<bool>,
-    pub disable_simultaneous_playback: Option<bool>,
-    pub no_embed_web: Option<bool>,
-}
-
-fn load_and_merge_app_config() -> Result<AppConfig> {
-    let mut config = load_and_parse_app_config()?;
-    let arguments = parse_arguments();
-
-    merge_option_with_args_and_env(&mut config.input_device, &arguments, "input-device");
-    merge_option_with_args_and_env(&mut config.output_device, &arguments, "output-device");
-    merge_option_with_args_and_env(&mut config.loopback_device, &arguments, "loopback-device");
-    merge_option_with_args_and_env(&mut config.stop_hotkey, &arguments, "stop-hotkey");
-    merge_option_with_args_and_env(&mut config.http_socket_addr, &arguments, "http-socket-addr");
-
-    merge_flag_with_args_and_env(&mut config.auto_loop_device, &arguments, "auto-loop-device");
-    merge_flag_with_args_and_env(&mut config.no_http_server, &arguments, "no-http-server");
-    merge_flag_with_args_and_env(&mut config.telegram, &arguments, "telegram");
-    merge_flag_with_args_and_env(&mut config.terminal_ui, &arguments, "terminal-ui");
-    merge_flag_with_args_and_env(&mut config.no_native_gui, &arguments, "no-native-gui");
-    merge_flag_with_args_and_env(&mut config.no_embed_web, &arguments, "no-embed-web");
-    merge_flag_with_args_and_env(&mut config.no_duplex_device, &arguments, "no-duplex-device");
-    merge_flag_with_args_and_env(
-        &mut config.print_possible_devices,
-        &arguments,
-        "print-possible-devices",
-    );
-    merge_flag_with_args_and_env(
-        &mut config.disable_simultaneous_playback,
-        &arguments,
-        "disable-simultaneous-playback",
-    );
-
-    Ok(config)
-}
-
-pub fn get_app_config() -> std::sync::Arc<AppConfig> {
-    GLOBAL_APP_CONFIG.read().clone()
-}
-
-pub fn load_app_config_from_disk() -> Result<()> {
-    *GLOBAL_APP_CONFIG.write() = std::sync::Arc::new(load_and_merge_app_config()?);
-    Ok(())
-}
-
-pub fn save_app_config_to_disk(config: &AppConfig) -> Result<()> {
-    save_app_config(config)?;
-    *GLOBAL_APP_CONFIG.write() = std::sync::Arc::new(config.clone());
-    Ok(())
-}
 
 pub fn get_soundboards() -> std::sync::Arc<SoundboardMap> {
     GLOBAL_SOUNDBOARD_MAP.read().clone()
@@ -121,24 +49,6 @@ pub fn get_soundboard(
         f.get(&id)
             .ok_or_else(|| anyhow!("no soundboard with specified id"))
     })
-}
-
-pub fn soundboard_position_sorter(a: &Option<usize>, b: &Option<usize>) -> std::cmp::Ordering {
-    match (a, b) {
-        (None, Some(_)) => std::cmp::Ordering::Greater,
-        (Some(_), None) => std::cmp::Ordering::Less,
-        (a, b) => a.cmp(b),
-    }
-}
-
-pub fn get_soundboard_ids_sorted_by_position() -> Vec<Ulid> {
-    let soundboards = get_soundboards();
-    let mut soundboard_positions = Vec::new();
-    for soundboard in soundboards.iter() {
-        soundboard_positions.push((soundboard.0, soundboard.1.get_position()))
-    }
-    soundboard_positions.sort_by(|a, b| soundboard_position_sorter(a.1, b.1));
-    soundboard_positions.iter().map(|x| *x.0).collect()
 }
 
 pub fn get_sound(soundboard_id: Ulid, sound_id: Ulid) -> Option<Sound> {
@@ -446,22 +356,12 @@ pub struct HeaderConfig {
     pub value: String,
 }
 
-fn load_and_parse_app_config() -> Result<AppConfig> {
-    let config_path = get_config_file_path().context("Failed to get config file path")?;
-
-    let toml_config: AppConfig = {
-        if let Some(config_path) = config_path.as_ref() {
-            let toml_str = fs::read_to_string(&config_path)
-                .with_context(|| format!("Failed to read_to_string {}", config_path.display()))?;
-            toml::from_str(&toml_str)
-                .with_context(|| format!("Failed to parse {}", config_path.display()))?
-        } else {
-            AppConfig::default()
-        }
-    };
-
-    info!("Loaded config file from {:?}", config_path);
-    Ok(toml_config)
+fn soundboard_position_sorter(a: &Option<usize>, b: &Option<usize>) -> std::cmp::Ordering {
+    match (a, b) {
+        (None, Some(_)) => std::cmp::Ordering::Greater,
+        (Some(_), None) => std::cmp::Ordering::Less,
+        (a, b) => a.cmp(b),
+    }
 }
 
 fn load_and_parse_soundboards() -> Result<SoundboardMap> {
@@ -497,6 +397,9 @@ fn load_and_parse_soundboards() -> Result<SoundboardMap> {
         ));
     }
 
+    soundboard_map
+        .sort_by(|_, a, _, b| soundboard_position_sorter(a.get_position(), b.get_position()));
+
     info!("Loaded soundboards from {}", soundboards_path.display());
     Ok(soundboard_map)
 }
@@ -505,15 +408,6 @@ fn load_soundboard_config(soundboard_path: &Path) -> Result<SoundboardConfig> {
     let toml_str = fs::read_to_string(&soundboard_path)?;
     let soundboard_config: SoundboardConfig = toml::from_str(&toml_str)?;
     Ok(soundboard_config)
-}
-
-fn save_app_config(config: &AppConfig) -> Result<()> {
-    let config_path = get_soundboards_path()?;
-
-    let pretty_string = toml::to_string_pretty(&config)?;
-    fs::write(&config_path, pretty_string)?;
-    info!("Saved config file at {}", &config_path.display());
-    Ok(())
 }
 
 fn check_soundboard_config_mutated_on_disk(path: &Path, last_hash: u64) -> Result<bool> {
@@ -567,79 +461,4 @@ fn save_soundboard_config(
         soundboard_config_path.to_str().unwrap()
     );
     Ok(())
-}
-
-fn parse_arguments() -> clap::ArgMatches {
-    let matches = App::new("soundboard")
-        .version(crate_version!())
-        .author(crate_authors!())
-        .about(crate_description!())
-        .arg(
-            Arg::with_name("input-device")
-                .short('i')
-                .long("input-device")
-                .about("Sets the input device to use")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("output-device")
-                .short('o')
-                .long("output-device")
-                .about("Sets the output device to use")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("loopback-device")
-                .short('l')
-                .long("loopback-device")
-                .about("Sets the loopback device to use")
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("print-possible-devices")
-                .short('P')
-                .long("print-possible-devices")
-                .about("Print possible devices"),
-        )
-        .arg(
-            Arg::with_name("no-embed-web")
-                .long("no-embed-web")
-                .about("Do not use embed web ui files"),
-        );
-
-    #[cfg(feature = "autoloop")]
-    let matches = matches.arg(
-        Arg::with_name("auto-loop-device")
-            .short('A')
-            .long("auto-loop-device")
-            .about("Automatically create PulseAudio Loopback Device"),
-    );
-
-    #[cfg(feature = "gui")]
-    let matches = matches.arg(
-        Arg::with_name("no-native-gui")
-            .long("no-native-gui")
-            .about("Disable native gui"),
-    );
-
-    #[cfg(feature = "terminal-ui")]
-    let matches = matches.arg(
-        Arg::with_name("terminal-ui")
-            .long("terminal-ui")
-            .about("Enable terminal-ui"),
-    );
-    #[cfg(feature = "http")]
-    let matches = matches.arg(
-        Arg::with_name("no-http-server")
-            .long("no-http-server")
-            .about("Disable http server api and web ui"),
-    );
-    #[cfg(feature = "http")]
-    let matches = matches.arg(
-        Arg::with_name("http-socket-addr")
-            .long("http-socket-addr")
-            .about("Specify the socket addr for http server")
-            .takes_value(true),
-    );
-    matches.get_matches()
 }
